+++ coordinator/migrations/2024-02-08-105459_change_user_email_to_contact_detail/down.sql
@@ -1,0 +1,4 @@
+ALTER TABLE
+    users RENAME COLUMN contact TO email;
+ALTER TABLE
+    users ADD COLUMN nostr TEXT;