+++ coordinator/src/orderbook/tests/registration_test.rs
@@ -22,7 +22,7 @@ async fn registered_user_is_stored_in_db() {
     let dummy_email = "dummy@user.com".to_string();
     let fcm_token = "just_a_token".to_string();
 
-    let user = user::upsert_email(&mut conn, dummy_pubkey, dummy_email.clone()).unwrap();
+    let user = user::upsert_user(&mut conn, dummy_pubkey, dummy_email.clone()).unwrap();
     assert!(user.id.is_some(), "Id should be filled in by diesel");
     user::login_user(&mut conn, dummy_pubkey, fcm_token.clone()).unwrap();
 
@@ -30,9 +30,8 @@ async fn registered_user_is_stored_in_db() {
     assert_eq!(users.len(), 1);
     // We started without the id, so we can't compare the whole user.
     assert_eq!(users.first().unwrap().pubkey, dummy_pubkey.to_string());
-    assert_eq!(users.first().unwrap().email, dummy_email);
+    assert_eq!(users.first().unwrap().contact, dummy_email);
     assert_eq!(users.first().unwrap().fcm_token, fcm_token);
-    assert!(users.first().unwrap().nostr.is_empty());
 }
 
 fn dummy_public_key() -> PublicKey {